+++ packages/smarthr-ui/src/components/InputFile/types.ts
@@ -2,6 +2,11 @@ import type { classNameGenerator } from './style'
 import type { ComponentPropsWithRef, ReactNode } from 'react'
 import type { VariantProps } from 'tailwind-variants'
 
+export type PreviewableObjectType = {
+  /** プレビューダイアログ内のFileViewerで検索機能を有効にするかどうか */
+  searchable?: boolean
+}
+
 type BaseProps = VariantProps<typeof classNameGenerator> & {
   /** フォームのラベル */
   label: ReactNode
@@ -10,7 +15,7 @@ type BaseProps = VariantProps<typeof classNameGenerator> & {
   /** ファイルリストを表示するかどうか */
   hasFileList?: boolean
   /** ファイルのプレビュー機能を有効にするかどうか */
-  previewable?: boolean
+  previewable?: boolean | PreviewableObjectType
   error?: boolean
   multiple?:
     | boolean
@@ -20,3 +25,6 @@ type BaseProps = VariantProps<typeof classNameGenerator> & {
       }
 }
 export type Props = BaseProps & Omit<ComponentPropsWithRef<'input'>, keyof BaseProps>
+export type LowerProps = Omit<Props, 'previewable'> & {
+  previewable: PreviewableObjectType | undefined
+}